backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiohttp
python-dotenv
